Move function definitions from header to cpp file
