+++ src/sensesp/types/position.cpp
@@ -1,0 +1,20 @@
+#include "position.h"
+
+#include "ArduinoJson.h"
+
+namespace sensesp {
+
+void convertFromJson(JsonVariantConst src, Position &dst) {
+  dst.longitude = src["longitude"].as<double>();
+  dst.latitude = src["latitude"].as<double>();
+
+  if (src.containsKey("altitude")) {
+    dst.altitude = src["altitude"].as<float>();
+  }
+}
+
+bool canConvertFromJson(JsonVariantConst src, const Position &) {
+  return src.containsKey("latitude") && src.containsKey("longitude");
+}
+
+}  // namespace sensesp
